requirements.txt
ipython==7.19.0
pytest==6.1.2
selenium==3.141.0
allure-pytest==2.8.31

